% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\name{data_doc}
\alias{data_doc}
\title{Create stub documentation for a data frame}
\usage{
data_doc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a character string containing a documentation stub
}
\description{
Create stub documentation for a data frame
}
\examples{

cat(data_doc(iris))
}
