% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\name{DATASET}
\alias{DATASET}
\alias{data_doc}
\title{Create stub documentation for a data frame}
\format{A CLASS with NROW observations on the following NCOL variables.
\itemize{
}}
\usage{
data(DATASET)
}
\arguments{
\item{data}{a data frame}
}
\value{
a character string containing a documentation stub
}
\description{
Create stub documentation for a data frame
}
\examples{
data_doc(iris) \%>\% cat()

}
