% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone.R
\name{extract_phone}
\alias{extract_phone}
\title{Extract phone numbers}
\usage{
extract_phone(x, area = NULL, na.return = NA)
}
\description{
Extract phone numbers
}
\examples{
p <- c(NA, "1234567", "1234567890", "(123) 456-7890", "123.456.7890", 12345678900)
extract_phone(p)
extract_phone(p, area=555)
extract_phone(p, area=555, na.return = "no phone number supplied")

}
