% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex.R
\name{reprex}
\alias{reprex}
\title{Reproducible Examples}
\usage{
reprex(x, ..., fig.width = 6, fig.height = 3, opts_chunk = list())
}
\arguments{
\item{x}{An expression. If not given, \code{reprex()} looks for code in
\code{input} or on the clipboard, in that order.}

\item{...}{Additional arguments passed on to \code{\link[reprex:reprex]{reprex::reprex()}}.}

\item{opts_chunk}{Named list. Optional
\href{{http://yihui.name/knitr/options/}{knitr chunk and package options}} that
are forwarded to \code{knitr::opts_chunk$set()} and \code{knitr::opts_knit$set()},
respectively.}
}
\description{
Create a reprex with default figure size set and easily adjusted.
}
\seealso{
\code{\link[reprex:reprex]{reprex::reprex()}}
}
