% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phone.R
\name{extract_phone}
\alias{extract_phone}
\title{Extract phone numbers}
\usage{
extract_phone(x, area = NULL)
}
\description{
Extract phone numbers
}
\examples{
p <- c("1234567", "1234567890", "(123) 456-7890", "123.456.7890")
extract_phone(p)
extract_phone(p, area=555)
}

