% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex.R
\name{reprex}
\alias{reprex}
\title{Reproducible Examples}
\usage{
reprex(x, ..., fig.width = 6, fig.height = 3, opts_chunk = list())
}
\arguments{
\item{x}{An expression. If not given, \code{reprex()} looks for code in
\code{input} or on the clipboard, in that order.}

\item{...}{Additional arguments passed on to \code{\link[reprex:reprex]{reprex::reprex()}}.}

\item{fig.width}{Width of figures.}

\item{fig.height}{Height of figures.}
}
\description{
Create a reprex with default figure size set and easily adjusted.
}
\seealso{
\code{\link[reprex:reprex]{reprex::reprex()}}
}
